scap: increase ores canary targets from 1 to 4
During the past deployments we realized that having only one
ORES node as canary limits a lot our ability to quickly check
if anything is right or not (due to how uwsgi and celery
interacts). Given the current total amount of nodes, 18, we could
bump the canary targets to something bigger like 2 hosts for
each datacenter.

Change-Id: I909b1e7b571aa3b6016075182e3280549f371729